Trie now using lowercase to search
The uppercase function was breaking the links, so it's now using lowercase ( Letter case is to make sure it can find words no matter how it's put in. Without it, the Trie will not be able to find words that don't exactly match how they were inserted. I.E. "Cat" is inserted into Trie and can be found as "Cat", but if you try to find "cat" or "CAT" it doesn't work)